Markdent v0.27 uses Specio for types
This updates the distribution to be in line with Markdent::Types from
Markdent v0.27 which has switched to using Specio instead of
MooseX::Types.

The changes required to do this are:

  - Surround the type with quotes to avoid bareword error
    (e.g., Str -> 'Str').
  - Changing the `isa` option value for `has` attributes to use
    `t('Type')`.
  - Changing the `isa` option key of MooseX::Params::Validate's
    `validated_*` to `type`.

A future change could be to replace MooseX::Params::Validate with
Params::ValidationCompiler like Markdent v0.27 does, but that can be
done at another time.

Fixes <https://github.com/keedi/Markdown-Pod/issues/8>.